pyproject.toml
[tool.poetry]
name = "github-file-watcher"
version = "0.1.0"
description = ""
authors = ["Jesse Evers <jesse@jesseevers.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sendgrid = "^6.8.0"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
